plugin: avoid plugins to be loaded twice
The original implementation will load plugins from both
(XDG_LOCAL_HOME|HOME)/.local/lib/deadbeef and configured path, the latter is
ussualy in $PREFIX/lib/deadbeef. However if the deadbeef is configured to be
installed in $HOME/.local, it may cause the each plugin will be load twice.
Despite that there must be a bug in a plugin which is not reentrant, it should
avoid this anyway.

Signed-off-by: Thynson <lanxingcan@gmail.com>